+++ src/main/java/com/example/somar6mais6.java
@@ -5,7 +5,7 @@ public class somar6mais6{
     public somar6mais6(){
 
     }
-    public static int somar() {
+    public static long somar() {
       
        return 6+6;
   }
